check if aikqverify exists before copying

+++ aikqverify.sh
@@ -5,9 +5,10 @@ OAT=$1
 # install aikqverify
 echo "install aikqverify"
 cd $HOME
-rm -rf aikqverify*
-cp $OAT/services/aikqverify/target/aikqverify-2.0.zip $HOME
-unzip aikqverify-2.0.zip
+if [ ! -e aikqverify-2.0 ]; then
+  cp $OAT/services/aikqverify/target/aikqverify-2.0.zip $HOME
+  unzip aikqverify-2.0.zip
+fi
 cd aikqverify-2.0
 make
 make install
